requirements.txt
langchain
langchain-community
openai
faiss-cpu
tiktoken
duckduckgo-search
python-dotenv
uvicorn
fastapi
pandas
